api/requirements.txt
fastapi>=0.100
uvicorn[standard]>=0.20
tensorflow==2.19.0
keras==3.10.0
numpy==2.0.2
pillow==11.3.0
python-multipart